% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_rates.R
\name{shadow_rates}
\alias{shadow_rates}
\title{Inverse Call Transformation}
\usage{
shadow_rates(.data, .group_col, .dbl_col, .eta = 0.015, .strike = 0)
}
\arguments{
\item{.data}{A tidy \code{tibble}.}

\item{.group_col}{The maturities column.}

\item{.dbl_col}{A column with interest rates.}

\item{.eta}{A \code{double} with the threshold between high and low interest
rates environments.}

\item{.strike}{A \code{double} with the call strike rate. Default's to zero.}
}
\value{
A tidy \code{tibble}.
}
\description{
Computes the "shadow rates" implied by the inverse call transformation.
}
\examples{
library(ggplot2)

data("us_yield_curve")

shadow <- shadow_rates(us_yield_curve, maturity, rates)

ggplot(shadow, aes(x = index, y = rates, color = maturity)) +
  geom_line()
}
