% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_yield_curve.R
\docType{data}
\name{us_yield_curve}
\alias{us_yield_curve}
\title{US Nominal Yield Curve}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55573 rows and 3 columns.
}
\usage{
us_yield_curve
}
\description{
US Nominal Yield Curve from 02-01-1990 up to 21-09-2021.
}
\keyword{datasets}
